+++ lib/Partials.php
@@ -7,7 +7,7 @@ namespace Wireframe;
  *
  * This class holds Partial objects and provides method access to rendering them with optional arguments.
  *
- * @version 0.4.0
+ * @version 0.5.0
  * @author Teppo Koivula <teppo@wireframe-framework.com>
  * @license Mozilla Public License v2.0 https://mozilla.org/MPL/2.0/
  */
@@ -36,13 +36,65 @@ class Partials extends \ProcessWire\WireArray {
     }
 
     /**
+     * Return the rendered value of named partial, or null if none found
+     *
+     * This is an alias for Partials::get($key)->render($args).
+     *
+     * @param string $key
+     * @param array $arguments
+     * @return string|null
+     */
+    public function render(string $key, array $arguments = []): ?string {
+        $partial = $this->get($key);
+        if ($partial instanceof Partial) {
+            return $partial->render($arguments);
+        }
+        return null;
+    }
+
+    /**
+     * Returns the value of the item at the given index, or null if not set.
+     *
+     * This method is overridden so that we can support getting partials recursively based on provided path.
+     *
+     * This version also supports a second argument, which is not derived from WireArray::get(). When a Partial matching
+     * provided path is found *and* an array has been provided as the second argument, said array is then passed to the
+     * Partial::render() method, and resulting markup is returned instead of the Partial object itself.
+     *
+     * @param int|string|array $key
+     * @param array $arguments Optional array of arguments
+     * @return WireData|Page|mixed|array|null Value of item requested, or null if it doesn't exist.
+     * @throws WireException
+     *
+     * @see \ProcessWire\WireArray::get()
+     */
+    public function get($key) {
+        $partial = null;
+        if (\is_string($key) && strpos($key, '/')) {
+            $subkeys = array_filter(explode('/', $key));
+            if (!empty($subkeys)) {
+                foreach ($subkeys as $subkey) {
+                    $partial = ($partial ?? $this)->data[$subkey] ?? null;
+                    if ($partial === null) break;
+                }
+                if (\func_num_args() > 1) {
+                    $arguments = \func_get_arg(1);
+                    if (\is_array($arguments)) {
+                        return $partial->render($arguments);
+                    }
+                }
+            }
+        }
+        return $partial ?? parent::get($key);
+    }
+
+    /**
      * Enables derefencing of WireArray elements in object notation.
      *
      * Note that unlike regular WireArray, we're going to prioritize local data items.
      *
      * @param int|string $property
-     * @return Partial|Partials|null Partial or Partials object or null if no matching item found
-     *
+     * @return Partial|Partials Partial or Partials object
      */
     public function __get($property) {
         $value = null;